requirements.txt
# Web Frameworks
fastapi==0.109.2
uvicorn[standard]==0.27.1
streamlit==1.31.1

# Database
sqlalchemy==2.0.25
supabase==2.3.4
asyncpg==0.29.0
alembic==1.13.1

# AI Vision
google-genai>=1.0.0

# HTTP & Parsing
httpx==0.26.0
aiohttp==3.9.3
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11

# Data Processing
pandas==2.2.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# Email
resend==0.7.2

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.1
Pillow==10.2.0
aiofiles==23.2.1

# Rate Limiting
aiolimiter==1.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.0
